assert: allow non-int returning functions to assert
Include GIT_ASSERT_EXT and GIT_ASSERT_ARG_EXT so that functions that do
not return int (more precisely, where `-1` would not be an error code)
can assert.

This allows functions that return, eg, NULL on an error code to do that
by passing `return NULL` as a second parameter to the GIT_ASSERT_EXT
functions.  In addition, functions that do not return a value (have a
void prototype) can make use of these functions, but there is no way to
signal to a user that an error occurred and care should be taken to
assert a valid state on subsequent calls into the library after such a
failure.